fix(朗读): 早退盯排版重试的 generation 参数跟齐 Long
prepareReadAloudGeneration 自多角色 TTS 起就是 Long(0L)，
a4df8467c 新增 scheduleEarlyExitRetry 时按 Int 收参，
CI compileAppReleaseKotlin 三处类型报错。本文件内该参数
只接收 generation，private 无外部调用方，放宽为 Long。

+++ app/src/main/java/io/legado/app/service/BaseReadAloudService.kt
@@ -818,7 +818,7 @@ abstract class BaseReadAloudService : BaseService(),
      * 后面排版时这一等就是好几秒（实测 ~4.7s 的哑声窗口）。早退后替这一轮盯着：
      * 本章分页一就绪立刻重起；起了新轮(generation 变化)就静默退出。
      */
-    private fun scheduleEarlyExitRetry(generation: Int) {
+    private fun scheduleEarlyExitRetry(generation: Long) {
         lifecycleScope.launch {
             repeat(40) {
                 delay(500)
